Added CLI interpreter
We added a full cli interpreter based on the yargs project.
We still have to decide how we will be expecting the credentials,
until now we are expecting the OTP to be passed on the terminal, but the
username and pswd I think should be kept in the env, but not assuming an env file.